backend/ml_services/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
Pillow==10.1.0
opencv-python==4.8.1.78
